added movie List and movie List item + Dependency Injection module
