+++ app/src/main/res/drawable/bg_details.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid
+        android:color="#7895B2"/>
+    <corners
+        android:topLeftRadius="30dp"
+        android:topRightRadius="30dp"/>
+</shape>
