+++ app/src/main/res/drawable/ic_round_date_range_24.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#000000"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M19,4h-1L18,3c0,-0.55 -0.45,-1 -1,-1s-1,0.45 -1,1v1L8,4L8,3c0,-0.55 -0.45,-1 -1,-1s-1,0.45 -1,1v1L5,4c-1.11,0 -1.99,0.9 -1.99,2L3,20c0,1.1 0.89,2 2,2h14c1.1,0 2,-0.9 2,-2L21,6c0,-1.1 -0.9,-2 -2,-2zM19,19c0,0.55 -0.45,1 -1,1L6,20c-0.55,0 -1,-0.45 -1,-1L5,9h14v10zM7,11h2v2L7,13zM11,11h2v2h-2zM15,11h2v2h-2z"/>
+</vector>
