+++ app/src/main/res/layout/fragment_movie_detail.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MovieDetailFragment">
+    tools:context=".ui.view.MovieDetailFragment">
 
     <TextView
         android:layout_width="match_parent"
