+++ app/src/main/res/layout/fragment_movie_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MovieListFragment">
+    tools:context=".ui.view.MovieListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
